Example project is updated
